+++ src/distinct.cpp
@@ -47,7 +47,7 @@ SEXP n_distinct_multi(List variables, bool na_rm = false) {
   }
 
   MultipleVectorVisitors visitors(variables);
-  SlicingIndex everything(0, visitors.nrows());
+  NaturalSlicingIndex everything(visitors.nrows());
   if (na_rm) {
     Count_Distinct_Narm<MultipleVectorVisitors> counter(visitors);
     return counter.process(everything);